requirements.txt
aiocache==0.11.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
cached==2020.12.3
charset-normalizer==2.1.1
decorator==5.1.1
disnake==2.7.0
dnspython==2.2.1
frozenlist==1.3.3
idna==3.4
isort==5.10.1
motor==3.1.1
multidict==6.0.2
pymongo==4.3.3
python-dotenv==0.21.0
yarl==1.8.1
